scripts_model/requirements.txt
matplotlib
scikit-learn
torchvision
seaborn
numpy
pandas